♻️ Refactor: 파일, 상수 분리, useLayoutEffect 적용
